+++ src/components/Breadcrumbs/Breadcrumbs.tsx
@@ -1,7 +1,6 @@
-import React from 'react'
 import { useLocation } from 'react-router-dom'
 
-const Breadcrumbs = () => {
+export const Breadcrumbs = () => {
 
     const location = useLocation()
     console.log(location)
@@ -10,6 +9,4 @@ const Breadcrumbs = () => {
             <h2>привет</h2>
         </div>
     )
-}
-
-export default Breadcrumbs
+};
