+++ src/components/Breadcrumbs/index.js
@@ -1,0 +1,1 @@
+export { Breadcrumbs } from "./Breadcrumbs"
