+++ src/components/CatalogItem/index.js
@@ -1,0 +1,1 @@
+export { CatalogItem } from "./CatalogItem"
