+++ src/components/Categories/index.js
@@ -1,0 +1,1 @@
+export { Categories } from "./Categories"
