+++ src/components/CategoriesItems/index.js
@@ -1,0 +1,1 @@
+export { CategoriesItems } from "./CategoriesItems"
