+++ src/components/Menu/index.js
@@ -1,0 +1,1 @@
+export { Menu } from "./Menu"
