+++ src/components/index.js
@@ -1,0 +1,6 @@
+export * from "./Breadcrumbs";
+export * from "./CatalogItem";
+export * from "./Categories";
+export * from "./CategoriesItems";
+export * from "./Menu";
+export * from "./PopularItems";
