+++ src/layout/Header/Header.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Link } from 'react-router-dom';
 import { useSelector } from "react-redux"
