+++ src/pages/AboutPage.tsx
@@ -1,4 +1,4 @@
-import Breadcrumbs from "../components/Breadcrumbs/Breadcrumbs";
+import { Breadcrumbs } from "../components"
 
 
 const AboutPage = () => {
